# Dev external occupancy map generation