--- a/.gitignore
+++ b/.gitignore
@@ -11,3 +11,4 @@ _*/
 
 /app
 data
+/.venv