--- a/exts/omni.ext.mobility_gen/omni/ext/mobility_gen/config.py
+++ b/exts/omni.ext.mobility_gen/omni/ext/mobility_gen/config.py
@@ -1,6 +1,16 @@
 import json
-from typing import Literal
+from typing import Literal, Optional, Tuple
 from dataclasses import dataclass, asdict
+
+
+@dataclass
+class OccupancyMapConfig:
+    z_min: Optional[float] = None
+    z_max: Optional[float] = None
+    cell_size: Optional[float] = None
+    origin: Optional[Tuple[float, float, float]] = None
+    lower_bound: Optional[Tuple[float, float, float]] = None
+    upper_bound: Optional[Tuple[float, float, float]] = None
 
 
 @dataclass
@@ -8,10 +18,13 @@ class Config:
     scenario_type: str
     robot_type: str
     scene_usd: str
+    occupancy_map_config: Optional[OccupancyMapConfig] = None
 
     def to_json(self):
         return json.dumps(asdict(self), indent=2)
     
     @staticmethod
     def from_json(data: str):
-        return Config(**json.loads(data))
+        data = json.loads(data)
+        data['occupancy_map_config'] = OccupancyMapConfig(**data['occupancy_map_config'])
+        return Config(**data)