--- a/exts/omni.ext.mobility_gen/omni/ext/mobility_gen/config.py
+++ b/exts/omni.ext.mobility_gen/omni/ext/mobility_gen/config.py
@@ -4,20 +4,10 @@ from dataclasses import dataclass, asdict
 
 
 @dataclass
-class OccupancyMapConfig:
-    prim_path: Optional[str] = None
-    cell_size: Optional[float] = None
-    origin: Optional[Tuple[float, float, float]] = None
-    lower_bound: Optional[Tuple[float, float, float]] = None
-    upper_bound: Optional[Tuple[float, float, float]] = None
-
-
-@dataclass
 class Config:
     scenario_type: str
     robot_type: str
     scene_usd: str
-    occupancy_map_config: Optional[OccupancyMapConfig] = None
 
     def to_json(self):
         return json.dumps(asdict(self), indent=2)
@@ -25,5 +15,8 @@ class Config:
     @staticmethod
     def from_json(data: str):
         data = json.loads(data)
-        data['occupancy_map_config'] = OccupancyMapConfig(**data['occupancy_map_config'])
-        return Config(**data)
+        return Config(
+            scenario_type=data['scenario_type'],
+            robot_type=data['robot_type'],
+            scene_usd=data['scene_usd']
+        )