--- a/exts/omni.ext.mobility_gen/omni/ext/mobility_gen/extension.py
+++ b/exts/omni.ext.mobility_gen/omni/ext/mobility_gen/extension.py
@@ -24,6 +24,7 @@ import glob
 import omni.ext
 import omni.ui as ui
 
+from isaacsim.core.utils.stage import open_stage
 from isaacsim.gui.components.ui_utils import (
     btn_builder,
     cb_builder,
@@ -33,15 +34,18 @@ from isaacsim.gui.components.ui_utils import (
     multi_btn_builder,
     xyz_builder,
 )
+import isaacsim.core.api.objects as objects
+import carb
 
 from omni.ext.mobility_gen.utils.global_utils import save_stage
 from omni.ext.mobility_gen.writer import Writer
 from omni.ext.mobility_gen.inputs import GamepadDriver, KeyboardDriver
 from omni.ext.mobility_gen.scenarios import SCENARIOS, Scenario
-from omni.ext.mobility_gen.utils.global_utils import get_world
+from omni.ext.mobility_gen.utils.global_utils import get_world, new_world, set_viewport_camera
 from omni.ext.mobility_gen.robots import ROBOTS
 from omni.ext.mobility_gen.config import Config, OccupancyMapConfig
 from omni.ext.mobility_gen.build import build_scenario_from_config
+from omni.ext.mobility_gen.occupancy_map import OccupancyMap
 
 
 if "MOBILITY_GEN_DATA" in os.environ:
@@ -87,9 +91,14 @@ class MobilityGenExtension(omni.ext.IExt):
             with ui.VStack():
                 with ui.VStack():
                     with ui.HStack():
-                        ui.Label("USD Path / URL")
+                        ui.Label("Stage")
                         self.scene_usd_field_string_model = ui.SimpleStringModel()
                         self.scene_usd_field = ui.StringField(model=self.scene_usd_field_string_model, height=25)
+
+                    with ui.HStack():
+                        ui.Label("Occupancy Map")
+                        self.omap_field_string_model = ui.SimpleStringModel()
+                        self.omap_field = ui.StringField(model=self.omap_field_string_model, height=25)
 
                     with ui.HStack():
                         ui.Label("Scenario Type")
@@ -99,17 +108,8 @@ class MobilityGenExtension(omni.ext.IExt):
                         ui.Label("Robot Type")
                         self.robot_combo_box = ui.ComboBox(0, *ROBOTS.names())
 
-                    with ui.VStack():
-                        self._omap_override = cb_builder(
-                            "Manual occupancy bounds",
-                            tooltip="If true, use manually specified occupancy map bounds",
-                            default_val=False
-                        )
-                        self._omap_origin = xyz_builder(label="Origin")
-                        self._omap_lower_bound = xyz_builder(label="Lower Bound")
-                        self._omap_upper_bound = xyz_builder(label="Upper Bound")
-                        
                     ui.Button("Build", clicked_fn=self.build_scenario)
+                    # ui.Button("Build", clicked_fn=self.build_scenario)
 
                 with ui.VStack():
                     self.recording_count_label = ui.Label("")
@@ -135,6 +135,7 @@ class MobilityGenExtension(omni.ext.IExt):
     def draw_occ_map(self):
         if self.scenario is not None:
             image = self.scenario.occupancy_map.ros_image().copy().convert("RGBA")
+            carb.log_warn(image)
             data = list(image.tobytes())
             self._occupancy_map_image_provider.set_bytes_data(data, [image.width, image.height])
             self._occ_map_frame.rebuild()
@@ -160,7 +161,8 @@ class MobilityGenExtension(omni.ext.IExt):
         self.keyboard.disconnect()
         self.gamepad.disconnect()
         world = get_world()
-        world.remove_physics_callback("scenario_physics", self.on_physics)
+        if world is not None:
+            world.remove_physics_callback("scenario_physics")
 
     def start_new_recording(self):
         recording_name = datetime.datetime.now().isoformat()
@@ -218,74 +220,64 @@ class MobilityGenExtension(omni.ext.IExt):
                 if self.step % 15 == 0:
                     self.recording_step_label.text = f"Current recording duration: {self.recording_time:.2f}s"
 
-    def get_omap_config(self, robot_type):
-        z_min = robot_type.occupancy_map_z_min
-        z_max = robot_type.occupancy_map_z_max
-        cell_size = robot_type.occupancy_map_cell_size
-        omap_config = OccupancyMapConfig(
-            origin=(
-                self._omap_origin[0].get_value_as_float(),
-                self._omap_origin[1].get_value_as_float(),
-                0.
-            ),
-            lower_bound=(
-                self._omap_lower_bound[0].get_value_as_float(),
-                self._omap_lower_bound[1].get_value_as_float(),
-                z_min
-            ),
-            upper_bound=(
-                self._omap_upper_bound[0].get_value_as_float(),
-                self._omap_upper_bound[1].get_value_as_float(),
-                z_max
-            ),
-            cell_size=cell_size
-        )
-
-        if not self._omap_override.get_value_as_bool():
-            omap_config.prim_path = "/World/scene"
-
-        return omap_config
     
-    def update_ui_from_config(self):
-        omap_cfg: OccupancyMapConfig = self.config.occupancy_map_config
-        for i in range(3):
-            self._omap_origin[i].set_value(omap_cfg.origin[i])
-        for i in range(3):
-            self._omap_lower_bound[i].set_value(omap_cfg.lower_bound[i])
-        for i in range(3):
-            self._omap_upper_bound[i].set_value(omap_cfg.upper_bound[i])
-
     def build_scenario(self):
 
         async def _build_scenario_async():
             
             self.clear_recording()
             self.clear_scenario()
+            self.disable_recording()
 
-            config = self.create_config()
-            robot_type = ROBOTS.get(config.robot_type)
-            config.occupancy_map_config = self.get_omap_config(robot_type)
+            # Get parameters from UI
+            scenario_type_str = list(SCENARIOS.names())[self.scenario_combo_box.model.get_item_value_model().get_value_as_int()]
+            robot_type_str = list(ROBOTS.names())[self.robot_combo_box.model.get_item_value_model().get_value_as_int()]
+            scene_usd_str = self.scene_usd_field_string_model.as_string
 
-            self.scenario, self.config = await build_scenario_from_config(config)
-
-            self.update_ui_from_config()
-
-            self.draw_occ_map()
+            robot_type = ROBOTS.get(robot_type_str)
+            scenario_type = SCENARIOS.get(scenario_type_str)
             
-            world = get_world()
-            await world.reset_async()
+            # Set config
+            self.config = Config(
+                scenario_type=scenario_type_str,
+                robot_type=robot_type_str,
+                scene_usd=scene_usd_str
+            )
 
-            self.scenario.reset()
+            occupancy_map = OccupancyMap.from_ros_yaml(self.omap_field_string_model.as_string)
 
-            world.add_physics_callback("scenario_physics", self.on_physics)
-
-            # cache stage
+            # Open stage
+            open_stage(scene_usd_str)
             self.cached_stage_path = os.path.join(tempfile.mkdtemp(), "stage.usd")
+
+            # Add ground plane
+            objects.GroundPlane("/World/ground_plane", visible=False)
+            
+            # Save the stage with ground plane
             save_stage(self.cached_stage_path)
 
-            if self.recording_enabled:
-                self.start_new_recording()
+            # Initialize world
+            world = new_world(physics_dt=robot_type.physics_dt)
+            await world.initialize_simulation_context_async()
+            
+            # Add robot
+            robot = robot_type.build("/World/robot")
+            
+            # Set the chase camera
+            chase_camera_path = robot.build_chase_camera()
+            set_viewport_camera(chase_camera_path)
+            
+            # Set the scenario
+            self.scenario = scenario_type.from_robot_occupancy_map(robot, occupancy_map)
 
-            # self.scenario.save(path)
+            # Draw the occupancy map
+            self.draw_occ_map()
+
+            # Run the scenario
+            await world.reset_async()
+            self.scenario.reset()
+            world.add_physics_callback("scenario_physics", self.on_physics)
+
+
 
         asyncio.ensure_future(_build_scenario_async())