- Adds performance optimizations related to get_world_pose().  Increases FPS from 20->60FPS for all scenarios
- Adds performance optimizations to path following scenario.  Increases FPS from 15->55FPS for this scenario
- Adds support for loading pre-generated Occupancy Map 
- Adds instructions for generating occupancy manually using existing Isaac Sim tool to README